requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic>=2.7.4
openai>=1.99.9
plotly==5.17.0
pandas==2.1.4
pytest==7.4.3
boto3==1.34.34
python-dotenv==1.0.0
# snowflake-connector-python==3.6.0  # Temporarily removed due to build issues on Windows
